Tournament creation + match gui
The tournament is tiered, with the winner of each stage going on to play against the other winners in the next stage. Added these UIs:
- Dialog form for choosing which players will be in the tournament
- Dialog form that shows matches one by one and lets the user choose the winner of each match. 

In a future commit I'll likely change this to show all matches in a stage at once. This would be useful since the matches might be completed in a different order than they're stored in memory.

Next I'm going to update the database tables, then do some UI improvements.